parse method names from thrift code instead
we need to return a list of methods from the raw thrift code sent in the
body. the point at which this API is used in the UI flow, we don't have
the thrift in the idl-registry and hence the the parse from file path
approach is not going to work.