Fix entries in build.yaml being order dependent
Because we are populating resolved dependencies inside the loop over
each glob pattern, we introduce an ordering dependency. If a module
that is matched by early glob pattern depends on something that is
matched by a later glob pattern, we will fail to find the module in
the list of resolved modules so far. Instead we should wait to
populate resolved dependencies until all of the glob patterns have
been evaluated.