yaml parser in PR#626 renders lines as dead code
initially when we moved from JSON -> YAML, there were some changes
needed on how to read Float but those are rendered useless as part of
PR# 626.

This dead code is causing coverage to drop. Since this code path is not
taken, a unit test can't be implemented either.

Deleting this deadcode here also helps coverage come back up
to 100% for ./runtime/* dir