support binary pass through response from clients
there are situations where an client response may not be able to conform
to a fixed schema as the payload may not be generated by a system that
is controlled by us. when talking to such a system, an artifical schema
may be created but its hard to keep it up to date causing other issues
hence its best to just allow binary pass through.

E.g., downstream is a 3rd party that returns a JSON blob and provides an
SDK to consume the JSON blob. In this case, the schema is negotiated
between the downstream 3rd party and the SDK provided by the 3rd party.