Reduce log-level of per-request messages
Move the "Finished ... Reqeuest..." messages from log-info level to log-debug level.

But a lot of the tests seem to rely on this being in the logs.
Therefore, having the test gateways run at debug-level logging.

There were a couple of null-ptr deref bugs that this change exposed (in the template)
Fixing those as well and piggybacking the changes