Allow only system failures to trigger circuit breaker (#667)
* Allow only system failures to trigger circuit breaker

When the downstream client call fails due to application errors, we should not count that towards
the circuit breaker threshold.  The systems are still behaving okay.

For tchannel, it is the "system errors"
For http downstream, we sorta count 5xx as system failures. This may not always be the case, but is there a better way?
For grpc downstream, all exceptions are system errors, so things should be okay.

* fixed an oops, return value overwriting error

* Code review comments + fix bug

* rename plus one bug fix