support for system wide whitelisting of default headers (#709)
* support for system wide whitelisting of default headers

we have a concept of default middleware, this is a same concept where we
don't need a whitelist of headers per endpoint but rather a system-wide
whitelist that will be always forwarded when available. These should be
used for more of control plane kind of headers rather than application
logic.