Extend TestGateway request creation with header multi-values (#737)
In order to use the changes in #733 in services that consume Zanzibar, we extend the TestGateway interface to mirror its existing MakeRequest method with one that creates a test request with all, disparate values for each header's key. We choose extending the interface instead of refactoring MakeRequest to preserve backwards compatibility with all of its many consumers.