Query param of type set implemented as slice (#742)
Query params that go into a Set datastruct will use a slice implementation

With this change any query param that needs to be unmarshaled into a set will have the set
implemented as a slice with unique elements. This is equivalent to the `(go.type = "slice")`
annotation.  The default implementation of a map with empty values is not useful for any workflow.